Resize fig for States above 90Percentile cases
Updated 90percentile calculation to work only on the last row of
cumulative sum data, which corresponds to the latest date, available
in the data set.

Resized fig and added one more row of plots.

Added the plot for states which fall in the 90th percentile and
above interms of cases they have, relative to other states.