Cov19In: Plot Tests/Cases
To check what Health Ministry/ICMR official told on news today,
sadly it shows that they were just telling a number in the wrong
light. The tests per positive cases, has been coming down as more
cases come up, but they tried to spin it has India is the best
in terms of this ratio.

It ignores the fact that we have a much bigger population and
Also one has to check if in most cases, were the cases identified
only after the cases reached the hospital OR was it pro-active
measures of reaching out to potential cases that identified them.