Cov19In: Added some informative prints
TODO: have to cleanup the data being imported from General data.json
file. There seems to be more than 1 set of data for some days.

Also for missing data, as this is a cumulative data set, I have to
use avg data wrt the day before and after the date with missing data