Cov19In: A better _fix_cumu logic
Instead of blindly putting the mid point between the prev and next
available data into a missing point (irrespective of the number of
inbetween missing points). Now it looks at the amount of gap (i.e
num of missing points between know values) and based on that decides
the value to put.

This should ensure a more smoother changing of data between the
known values.