Cov19In: Fix a issue with indexing into numpy ndarrays
The output of argwhere can't be used directly to index into the
array to get the corresponding elements. One is required to convert
each row (which corresponds to a given n-dimensional address)
in the output of argwhere into tuple before indexing.

If one doesnt do this, then the list of numbers will be used to
select all rows which correspond to those numbers, instead of
selecting the specific element of the array which corresponds to
the given n-dimensional address.

Ex: if 3-dimension address is [1, 5, 3], instead of selecting the
specific element at ndarray[1, 5, 3], it will select ndarray[1],
ndarray[5] and ndarray[3].

NOTE: Have purposefully left the prints and code used for debugging/
confirming still in there. WIll clean it up next.