Cov19In: Added inset boxplot with log scale for cur boxplot
The linear boxplot doesnt show details of states with low incidence
of cases. So this log boxplot inset added to see the details related
to such states.

The inset boxplot is made translucent so that the original boxplot
can still be seen properly, even thou the inset boxplot overlaps it.