Generate png files, make inset face transparent
There were issue with converting the generated svg files to png
using both Imagemagick and Graphicsmagick. So generate png from
with in pyplot.

Make the inset face/background transparent for the rel-improving?
and rel-worsening?, so that one can see the parts of main plot
which is overlapped by the insets.