EuWorldDataSrc: Conv got data for use - part1
The EuWorld data is in a verbose format, as well as contains
additional data, not currently used by my logic, so it needs
to be converted into the simpler format also used by Cov19In
DataSrc, so that common logic can be used to analyse both data
sets.

Have started the conversion process in this commit.

Date conversion to float handled.

TODO: however noticied that the data contains the csv delimiter
"," also as part of some fields which are protected using '"'.
However numpy.genfromtxt doesnt handle it, so have to clean it
up.