AnalPlot: Check for None using type
As numpy arrays override == operator, so one cant check if some
given variable is a None or has some data like a list or numpy
array or so.

So have changed to explicitly check the type to see if it matchs
None type or not.

NOTE: THe previous simple check will work for normal python types
like string or list or so, but not for numpy arrays.