ColHdr ndarray, Skip date & total cols wrt ColHdr
For working with special indexing on Column Header list so that the
selected cols can be returned, it requires to be of ndarray type
and not normal list. So added logic to AnalPlot.set_raw to do the
conversion to ndarray if required.

In the program, even thou the data set passed to set_raw was skipping
the date and total cols, the column header passed wasn't skipping
these. This oversight has been corrected now.