AnalPlot: by default inclusive colsels_percentile
Previously the percentiles specified in the range of percentiles
to use for selecting cols, wasnt inclusive. But have changed the
logic now such that the specified percentiles of the percentile
range boundry are inclusive.

If one wants the old behaviour, where the boundry percentiles
of the percentile range were not included as part of logic for
selecting cols, then pass bSelInclusive to False.