AnalPlot: calc reqd percentile range given topN or botN
If we are interesting in looking at data related to the top N
or bottom N number of regions or cols in the data set, then
instead of we having to calculate what the corresponding percentile
range should be for selcols_percentile, now it will calculate the
range on its own, if either topN or botN is specified.

So also it returns the percentile range of selected columns now.

Main program updated make use of the above new selcols_percentile.

SO it no longer tries to find the percentile range related to
top 15 regions, on its own. It lets the selcols_percenitle do the
calc on its own.

Also simplified the call to np.percentile. As I can send the single
selected row as a 1d array rather than the prev almost dummy 2d
array(which is actually a 1d array in a 2d cloth).