DataSrc: Fix Missing Part 2
Added a new function fix_missing, to which one passes the arguments
as a dictionary. It inturn calls a specific logic as specified.

DataSrc.load_data updated to use the same.

Cov19In and EUWorld load_data updated such that if user has not
specified any fixMissing argument, then it uses fix_missing_value
to replace NAN with 0.