EUWorldDS: Save as csv with int data - part 2
If fmt not specified, then automaticaly saved as float(eng notation),
irrespective of the datatype of passed ndarray?!.

So using fmt to save as int data, because the data in the array is
only data. NOTE: Missing data is replaced with 0 and not as local
mean or some other logic, so int is valid.