AnalPlot: _get_datakeys, set_raw
As one requires to generate the dictionary keys related to data,
its row header and its col header often, so added a helper func.

set_raw updated to use the same.

Also made set_raw take a dataName, which will be the base name used
for the dictionary keys corresponding to it. In turn the same is set
to "raw" as default, so that the old behaviour is still available
as default.

However one can now store different sets of data, ColHdr & RowHdr
by giving different names to them.