AnalPlot: Rename dataSel and dataName to dataKey
All the set & calculated data is stored in a dictionary by AnalPlot,
inturn dataSel is used to refer to it in most of the functions. And
dataName is used to refer to it in set_raw.

Now I have renamed it to dataKey. This also matchs the _get_datakeys
helper func.