AnalPlot: new_dataset, comment for some funcs
Moved the logic which clears and sets up the AnalPlot instance
for new dataset into a new function called new_dataset, from
the set_raw func.

So the main program updated to call new_dataset as required.

NOTE: THis is a continuation of making set_raw into a generic func
which can be used to store multiple raw data sets into AnalPlot
instance.