AnalPlot: get_data_selective, plotXY
Added the get_data_selective helper.

Inturn same used by plot, boxplot, as well as the new plotXY.

Added a plotXY, which plots the selected list of cols along a
XY plot, where values of those cols in one of the specified
dataKey's acts as the value for X axis and the values of these
cols in the other specified dataKey  acts as value for Y axis
for the XY plot.