AnalPlot:setFigBoundries incaseTightlayoutBailsOut
THere seems to be some issue with the logic used by pyplot tight
layout logic, sometimes it bails itself out.

So have implemented a simple logic in my subplots helper in AnalPlot,
which calcs the values related to margins and spacing between plots
to allow the usual parapernalia around plots, using the default font
in my system.

NOTE: Currently it doesnt cross check the font that will be used
and its avg width and height per char, so if the font changes the
logic may require adjusting. Also it is a static logic and doesnt
depend on the actual details of the things put around the plots
in any specific situation.