AnalPlot: get_data which creates calc data if reqd
Old get_data is now called _get_data.

The new get_data checks if the data requested is already there
or not. If the data is not there, it checks if the data can be
calculated from the data already available. If possible, then
it creates the data by calling required calc_xyz functions.

As part of it, it moves towards the base data, across multiple
steps if required and recreates all calculated intermediate datas
in order to generate the actually requested data.

NOTE: When assigning function pointers, the func should be already
declared.