AnalPlot: _textxy a saner one
Reverted to log10 as required and noted before.

The conflict detection also uses a fraction of the current value
being checked against.

The random delta being added had a bug, in that uniform requires
both low and high, if only one arg given, it wont consider it has
the high, so that was also messing the thinking based on test runs.