AnalPlot: help _textxy use prev updates when deciding new
Previously each new text plot xy position was decide in its own
universe (to itself), without worrying about other texts.

Now when ever a new text xy position is decided, its updated into
the current list of text xy positions. So that when the remaining
text xy positions are decided in future within that plot, it will
cross check that this new text is not overlapping with this updated
list of text positions.

This ensures that if there is a bunch of near by text elements, and
if moving one of them away, avoids overlapping of the texts, then
the other texts will no longer be moved around.

TODO: There is still one more thing to handle, i.e if based on current
list of text elements, if it is decided that a given text has to be moved,
currently the new position calculated for it is not crosschecked against,
the list of text positions, so it can end up overlapping a previously
plotted text.

i.e Currently whether a text needs to be moved or not uses the uptodate
list of text positions, but when something needs to be moved, the new
position calculated is not crosschecked against the list of text
positions.

Also random turned out to be bit too random and not helpful. And
blind rotating also true to probability didnt help much, as we
will be going through the list of texts to plot not in any specific
order, so the blind rotating logic may try out adjacent text elements
after trying other elements inbetween, so the rotating doesnt necessarily
help avoid overlapping of adjacent elements.