AnalPlot:newxy_rotate: Better debug print, ALSO REALISED
Also realised that in textxy_super, I was passing the last got x,y
back to _textxy not only to get a spiraling motion, but also to check,
if the new location overlaps with any text or not. But by avoiding
that I have broken the check logic.

Will be fixing this oversight next.