AnalPlot:textxy: Split to overlapcheck+update, proper super
The _textxy logic split into two funcs

a) check overlap func

b) if overlap then get new x,y func

In turn _textxy_super updated to now cross check if the new x,y
got overlaps any other text in the list given to it or not. This
is a fix for a oversight in a previous update.

The logic works now. Can it be made more beautiful needs to be
thought of, but otherwise ok.

Also in the inset the movement seems to be larger, need to cross
check this still.