AnalPlot:textxy.newxy_rot: Very slower spiraling
Made the spiraling out based finding of new x,y for overlapping
text in newxy_rotate lot more slower now.

Inturn to accomodate the same the num of attempts in textxy_super
increased to some extent. Doesnt fully compoensate for the new
slower spiral, but still should be good enough in general.

NOTE: The spiralling is bit more slower than what is required, but
just to be overtly sure for now.