AnalPlot: custom textxy_spread; Main: tighter inset plotxy
Added a custom mode to the textxy_spread logic, so that one can
specify ones own multiplier to apply to the current lever values.

Inturn the same is used to make the inset plotxy in the main program
tighter so that it appears ok (Need to finetune the textxy overlap
avoidance logic to be more saner for such cases also later).