AnalPlot:test_plotxy_rect helper func
Added a helper function to plot the 4 edge vertices of a rectangle
given to it and inturn use the same to plot the rectangular overlap
cross check area of each text being plotted by textxy.

This helps verify that the rectangular overlap check areas are ok
or not.

It also allows 2 given messages to be plotted next to the rectangle