AnalPlot: test_plotxy_rect helper for controlled use
Renamed the base code as _test_plotxy_rect and inturn added a
helper logic on top, which takes a id as a additional parameter.

It inturn checks if the AnalPlot instance has a dictionary named
dTestPlotXYRect, and if the current id passed to it is part of
this dict. If so, it checks if the use counter related to this id
in the dictionary is greater than the max use counter in the same
dict. If so it ignores the call, else it plots the rectangle and
increments the use counter.

plotxy also updated to setup the dictionary used by this logic with
id's of interest, which are the data set column names.  And textxy
passes the column header name passed to it to the test_plotxy_rect
helper func.