AnalPlot: axis_adjust better logic
axis_adjust done between movies rajamma@yahoo and megasandesham ;-)

This should handle most cases in a ok way.

Had looked at this path to see, if the range checked by textxy
logic to avoid text overlapping was uniform across different
plots/plot_sizes or not,

i.e by checking if the range of axis and the data range mapping
on to the axis used during calc and later during plot needing to
be fixed, but this doesnt seem to have helped.

This was required, bcas, I found the rect dimension/size changing
between different plots. Need to look at this further later.