AnalPlot:textxy: fixed range calc oversight, oops
Identified the reason why the overlap area check of textxy was not
consistent across different plots.

I was calculating the xRange and yRange wrt x-axis and y-axis data
wrongly for the log scale plots.

In overenthu to minimize the code, previously I was stupidly doing
log10(Max-Min) instead of log10(Max) - log10(Min), thus the xRange
and yRange was invalid for plots with log scale.

The plotxy_rect which I had added in the last commit relating to a
rect between 1/4th of x and y axis to 3/4th of x and y axis, to help
try identify the issue, helped find the issue readily, because the
rect was plotting in the wrong place, by a huge margin. So some
other possible reasons, which I had in mind as to why overlap check
could be messing up like aspect ratio, plot changing dimension when
saving file or any such corner case could no longer be contenders.
And that gave the required trigger and enthu to relook at things
properly.

Also identified that the pixel dimensions specified for text chars
is not perfect, I require to increase the width bit more.