AnalPlot:axis_adjust handle if 0 min for log plots
Now axis_adjust takes care to ensure that xMin and yMin for a log
plot is not 0. If it finds that min is 0, then it replaces it with
the next minimalist value in the array of values, ie the value
immidiately larger than 0 in the array, but smaller than any other
values.