AnalPlot:axis_adjust a better version.
Now in case the xMin or yMin is found to be 0, then instead of
blindly replacing it with the next smallest value, a suitable
value derived from the next smallest value is used. This should
ensure that the data points are not exactly at the edge/boundry
of the plots, but instead slightly within the plot.

Readme is also updated.

Also disable textxy debug mode.