P1: scale diffData b4 plotxy, verbose NotImplementedError
As the diff data wrt cases/day can have either +ve or -ve values
in it, so one cant plotxy the same in log scale properly, if the
selected cols (i.e regions) contain -ve values.

To overcome this now I scale the diff data, before trying to plotxy
it.

TODO: the calc_scale currently prefixes the inDataKey to the passed
outDataKey before using it has the actual outDataKey, need to make
this more controllable.

Updated the get_data's NotImplementedError info to be more verbose.
Also this fixes a slight oversight in there.