AnalPlot: WorkAround a bug in PyPlot.arrow
If the base x,y value of arrow is less than 1 and inturn very near
to 0 (say like 0.0005), and the delta x and delta y passed to arrow
is 0.0, then atleast for log scale plot, the arrow plotted can be
wrong (i.e instead of no arrow because dx and dy is 0.0, it may
paint a long arrow).

This is worked around by checking if delta x and delta y being
passed to arrow func is 0 or not. If both are 0, dont call arrow.