AnalPlot:calc_scale: Fix wrong calc issue - p1
When doing scaling of cases/day.diff data and inturn plotting, it
was noticed that some data points were plotting in the wrong place,
as calc_scale was just implemented and not verified, so this branch
and this set of commits is to identify and fix the issue.

To help with same, the normal plot of diff data is now diff.movavgT2
instead of diff.movavgT3, so that this diff data corresponds to the
same one used in plotxy.

Also added initial skeleton of a print_data_selective, logic to help
with this debugging.

One possibility that comes to mind is, may be some precision loss is
creating this issue. However I need to cross check this.  Or maybe
some oversight, as I just implemented the calc_scale in oneshot
without revisiting/cross checking the code which i wrote. And things
have caught up with me now, forcing this cross check.

Equally it could be some issue with scale diff data logic as I just
implemented a basic experimental code for it earlier today and havent
thought through it fully yet.