Main: Scale diff data and plotxy cross check - p2
Printed 2 regions, which I noticed, were changing position in the
plot based on whether raw diff data is printed or scaled diff data
is printed.

Looking at it, as well as thinking about whats occuring bit more,
I think the problem has to do with me scaling the diff data relative
to the respective regions, instead of across the regions.

For the plotxy case, I am comparing data across the regions, so when
scaling and plotting, I require to scale the data from across the
regions for the value in the last row in each region.

I think fixing this will fix the issue.