AnalPlot:calc_scale: On 2D data for axis=0 or 1
Now calc_scale can scale data by looking at neighbours along either
the rows or the cols in the dataset. So the dataset can be 2D.

Previously it was always scaling data by looking at data in each
column in the dataset. i.e scale data along each column, by looking
at the charactersitic of the data in each given column. i.e this
is equivalent to axis=0, in the new logic.

TODO: The logic which I have added for if axis==1 condition, is a
generic logic, which can work for either axis=0 or axis=1. Except
for the logic at the end wrt reshaping the inMin, inMax, outMin and
outMax arrays. Will simplify and update the flow later.