Main: Kulja Simsim, Scale Diff plotxy flying
As noted earlier, the reason the plotxy of cases/day.diff scaled
data was messed up a bit was because, in my hurry to implement
the scaled data plotting, I was scaling the data within each region
instead of across regions.

As I was plotting xy data to help compare data of one region wrt
other regions, across 2 parameters (one of them on x and other on
y). I had to scale the data relative to same data of other regions.

Now I have fixed this, so the plotxy of scaled diff data is fine
now.