AnalPlot: Fix tick params for log scale also
For the tick params to apply to both linear and log scale or rather
both major and minor ticks, one requires to set which="both" to the
required plots. So same is done now for plotxy.

TODO: have to propogate this change/update to other plot types also.

Also have resequenced the test runs in hkvc-run.sh, so that the
saved image wrt auto fetched data corresponds to the scaled diff
data.