AnalPlot:movavg made generic to account all columns
Most probably bcas, I was using the 0th column (of python/c notation,
which is the 1st column literally speaking) to maintain the date info
in the original covid-19 related datasets in the previous logic.

Or because I was using the 1st column to maintain total for each row,
rather because one of the data sources was following that format. So
even that could have lead to this corner case, provided I was keeping
date column out and retaining this totals column still and didnt want
to operate on it or ...

I potentially carried forward one of the above assumptions into the
movavg calculations or else I was dreaming at that time, no idea now.
But the end result was that the 0th column was getting forced into
all 0s.

The new usage/test sample code additions into analplot, reminded me
of this fact, based on its outputs (or else uniform distribution
wouldn't have been that random, now come on that cant be the case,
unless i was drunk ;-). As there was always a line perfectly in line
with the y = 0 i.e x-axis at origin of sorts.

So have fixed the movavg(_ex) code to handle 0th column also properly.

TODO: At a later stage have to make many of the calcs flexible from
2D datasets perspective by allowing them to support a axis argument.
For now most do their calculation along each given column i.e axis=0,
because my covid dataset is structured for that kind of calc in most
cases.

Also if I remember correctly now, when I implemented this built on
generic classes like datasrc, analplot based new covid19 toolkit prg,
I took care of moving this date column into the row header field of
the AnalPlot dataset structure. And inturn the totals column was
discarded when moving data in datasrc into analplot, as one can
calculate the same from the available data, and also because I was
not setting the totals column data for EUWorld dataset, if I remember
correctly now. So in that sense this bypassing of 0th column was not
required for a long time now.

This also means the 1st region in the respective datasets, was noping
itself wrt movavg. But otherwise everything else was fine, because
analplot caches all calculated data in its data dictionary by default,
and reuses the same, when a subset of columns are requested. So it was
not trying to recalc the movavgs for the subset of given data columns,
else this would have nop'd/nullified/zerod the movavg data for 0th
column in each subset of cols, and I would have noticed this a lot
earlier. But the reuse cached data flow, avoided this possibility, but
equally protected from getting/working on invalid data, except for the
0th col region in the original dataset.

A long story short, things are fine now, hopefully ;-)