Main: diff.movavgT3Plot easy4Users and cleanup
As diff data has too much entropy in it, even a 2 level movavging
doesnt make it easy for the eye, so have gone back to the original
3 level movavg plot. NOTE: Either way I had changed to a 2 level
movavg plot to previously cross check if the scaled diff data plots
had any issues in them or not by mapping the relative hierarchy of
the diff.movavgT2 plot curves to the relative positions of the data
points/regions in the inset plotxy of scaled diff data. As that
issue/mix-up has been fixed now, so this rawer movavgT2 is no longer
required.

Removed the commented out code wrt testing validity of scaled diff
data and its plot, which was printing the underlying data, before
and after scaling.

Also isolated running the plot_movavgs test code into a test flag.