AnalPlot:textxy_checkoverlap ref rect generified
the code which I had added to the textxy_checkoverlap logic, to
verify whether its geometric calculations are anchored in reality
or not by plotting a rectangle which spaned between 1/4th to 3/4th
of the plot window, by using the results of its geometric calcs,
has been made generic now. PReviously the few additional code
required for plotting this ancoring/reference rect was usable only
for a log scaled plot and not normal plot. WHich is fixed now.

If I didnt do it now, I would have forgotten about it with time.
And in future if one were to test something with a linear plot
window, then results would have been strange and confused the hell
out then.