git submodules used updated to their latest upstream version
Used >> git submodule update --remote << to sync/update the copy of
the submodules used by this project to the latest commit available
from my github repository for these submodules.

This updated version of pyuno_toolkit waits for libreoffice to become
available in a more intelligent way. So even when the program is run
the first time, chances are it will connect to libreoffice successfully
even if it takes more time.