AnalPlot:calc_ functions updated to support outDataKey
Now all the calc_??? functions support one to specify a outDataKey.

If one is specified, then the calculated data will be stored inside
the AnalPlot instance under this given outDataKey.

If one doesnt specify any outDataKey, then it is assumed to be __AUTO__
by default and inturn mapped to the core functionality related name,
which is also part of the calc_??? function name. Thus it duplicates
the old behavior. For example calc_rel2mean will use rel2mean as the
key id to suffix to the input data's datakey while deciding on its
outDataKey.

To help achieve the above logic, a helper func _outdatakey has been
added now.