AnalPlot:calc_scale inDataKey renamed to dataKey
This is to ensure all calc functions follow the same convention.

Also other functions also call the dataKey they use for identifying
the data to operate on as dataKey itself and not inDataKey.

However the _outdatakey helper function, names the input datakey,
it uses to generate the out datakey as inDataKey.