AnalPlot:DataOpsChainingP4: Added skeleton for funcs with args
If the data op func name is part of the dCalcFuncsWithArgs, then
it extracts the arguments if any specified in the DataOpsChaining
based datakey and inturn calls the helper function defined to
call that data ops function.

This helper function should extract the arguments from the given
argument list and inturn call the actual data operation func with
the arguments.