AnalPlot:DataOpsChainingP7: handle custom outDataKeys
DataOpsChaining for DataOps/funcs with arguments involves dataKeys
which are nonstandard because they have to mention the arguments and
their values in the dataKey.

THis inturn means that DataOpsChaining call helpers for data ops funcs
need to be passed the (out)dataKey being generated also.

This commit handles this.

Also there was one place in the usage/test sample code, where I had
forgotten to change from the old explicit called scale func and its
custom outDataKey to the new dataKey corresponding to DataOpsChaining
for funcs with arguments. I have fixed this also.