AnalPlot:DataOpChainP9: betterCodeflow4 callHelpers
call_calc_scale updated with a new code flow, which makes the logic
more powerful, flexible, robust and simple at the same time.

Now one can have aliases for the arguments like axis and A for axis.

If a specified argument is not understood, then it is flagged to the
user, by printing info about same.

Defaults for the arguments are captured cleanly at one location at
the beginning of the flow.