Main:rel2sum moved into 1st row inset, reduced rows
The plots simplified. The number of rows reduced by 1 by moving
the rel2sum plots from their own row to be the inset in the 1st
row.

The old raw data which was plotted in the inset is no longer plotted.
Which is ok at one level, because the 1st row already contains the
log scale plot of movavgd raw data.