Readme updated, helper script uses cur date
For the user selected region geoId based run of the program, now
the helper script auto identifies the current date to use. However
if the program had not been run previously to successfully fetch
and generate the processed csv files for the data sources, it will
fail. The helper script runs the main program without any argument
to help with this automatically once, but if it is run too early
in the day, the EUWorld data may not yet be available.

The Readme updated with bit more info about the python modules of
the program, which can be used on their own by other programs, if
required.