AnalPlot: movavg_ex int indexDelta, test_data_simple
As arithematic operations by default generate decimal values, even
for integer data, so the indexDelta which was calculated in last
commit in movavg_ex can't be directly used to adjust the headers.

So now indexDelta has been type casted into int, so that it can be
used to index into the existing header to generate the new reduced
index.

Also a new test_data_simple func has been added to help generate
a very simple dataset for testing.