AnalPlot: fix oversight wrt set_raw missing row/col header handling
If a variable is a numpy array, then we can't directly compare it
against None, so have updated the logic to find the type and then
check the type, to see if the passed row or col header is None or
not.

This was a oversight, but also captures the above quirk compared to
other standard types, indirectly.