Main: As movavg doesnt shave-off rowhdr, so plot xticks
In previous commits I had modified movavg logic so that they don't
blindly reset the row or col header to numeric counter like values,
instead they adjust the headers to match the reduced data set.

So the previous avoiding of custom xticks for cases/day>rel2sum>movavg
plot, has been reverted, so that it shows the date info in the plot
now.