AnalPlot: plotxy doc, arrow in neutral color
Have updated the plotxy documentation to be hopefully more clearer.

Have changed the arrow color from red to blue, which is more neutral
in nature. So that people dont interpret the end plot arrows to contain
any meaning, which is not there in the first place. The arrows just
indicate the text belongs to which point in the plot.

Later once I add showing the text or the marker in color, In the simple
case of indicating two groups, normally red and green can be used to
indicate something is towards one or the other end (i.e in one or the
other group), and blue for arrows remains neutral, in general.