AnalPlot:group_simple:p2:group based on percentile range
Instead of selecting the top N cols of the dataset, now the logic
has been udpated to select a group of columns based on the percentile
range specified.

This makes the logic more flexible, because previously one could only
select N of the top cols, based on a given criteria. Now instead one
can either select from the top of the group or the bottom of the group
or from anywhere in between.

Also fixed a oversight in last commit