AnalPlot:group_simple:p3:return numericBased ColSelector also
Now group_simple not only returns the list of selected cols' colHdr
values for selCols which fall in the percentile range. It will also
return a selection list which is numeric, and it contains 1 for the
selCols which fall in the given percentile range and -1 for selCols,
which fall outside the percentile range.

i.e selColsInPercentileRange related positions have 1 and
selColsOutsidePercentileRange related positions have -1

Also the Main Covid19 analtoolkit program has been updated to use this
new group_simple to highlight regions which are having cases/day changing
faster for the worse in the last few days (or put differently cases/day
changing relatively speaking in a worse way compared to others in the last
few days).

This means that if the cases/day is increasing in certain regions faster
compared to others, then it will be highlighted. While regions which may
be having much higher cases/day in absolute terms, but then if these
numbers have been reducing, then it will not be highlighted.

While if we are in a phase, where cases/day is reducing in all regions,
then the regions where the cases/day is reducing slowest will be
highlighted, compared to other regions.