AnalPlot:group_simple_percentiles cleaned up
Now it not only returns a list containing column header values of the
percentile based selected cols, but also a list containing the column
header values of the non selected cols.

NOTE: There is still the list of number based selColsP, which groups
the cols using the numbers -1 (not selected) and 1 (selected).

To avoid any ambiguity or confusion between selCols passed by user to
decide the subset of data one should work with. And the selCols selected
from with in the user specified selCols, based on the percentile range
given. The selCols created based on percentile range is named selColsP.

The documentation has also been updated a bit.