AnalPlot:plotxy:fixColorControl Val2ColMapthroulimits
If there was more than 1 entry in the color control limit array,
the logic would have always mapped all colorControl values lesser
than the last color control limit entry to the last but one color.

NOTE: Rather than just color, these control the marker style which
includes color and type of marker to use etc.

This bug with the logic has been fixed. I had just put the code in
there and not yet had time to check it till now, so good that I
found and realised this oversight with the 1st suitable test run
itself.

Also fix a oversight in the test code in the call to plotxy after
groupsimple_neighbours, wrt the color control limits array passed.