AnalPlot:GSNeighbours: plot to help debug
Added a optional axes argument to gs_neighbours. If provided, it
will use it to plot the intermediate results of its logic, to help
debug it.

For some randomly generated data, the logic seems to be doing what
its supposed to do to a great extent. Voyager and direct brain dump
in one go without any additional cross check seems to have gone well,
captian;)