AnalPlot:GSNeighbours: Cleanup arguments
Have removed some of the additional arguments which I had initially
added to the function, but am not using in the simple neighbour
logic for now, to keep things simple ;-).

Also the ratio of total data bounding rectangle diagonal distance,
which is used to decide whether two pointsOfInterest are neighbours
or not is now controllable by the caller.

Also updated the documentation of the func