AnalPlot:GSNeighbours: more inteli collasing of localcenters
Now the local centers identified are cross-checked to see if they
are near enough to merge into a new local center. And this test
is done, till we no longer get new collased/merged local centers.

However this also has the side-effect that, local centers can move
away from the edges, through merging. And this can in some cases,
lead to the edge point(s) becoming nearer and inturn assigned to a
previously unrelated but relatively speaking near in a way local center.