AnalPlot:GSNeighbours:Ensure stable minimal local centers reached
Now the logic for grouping based on whether they are neighbours or
not (based on distance) not only checks to see if a minimal set of
local centers have been identified, but also checks to see if that
minimal set is stable/consistent.

This helps collase things further down into a even smaller set of
local centers for grouping, in some cases.

TODO: Later have to think can this lead to a fixed size jumping set
of centers by any chance. On a very casual over the tides / waves
thinking of few seconds feel less likely, but need to think bit more
later. But either way, one can always exit with a simple same set size
repeat counter threshold logic.