AnalPlot:GSNeighboursEx: Try to control groupNumber in given N tries
It tries to get GSN to group things into the required number of groups
by trying to increase or decrease the diagRatio.

Inturn it attempts this only for a given finite amount of tries.

NOTE: It doesnt try to accelerate or declerate the diagRatio delta,
based on currently achieved numOfGroups and tries remaining.