AnalPlot:GSNeighboursEx:Fix few oversights.
Also note that the plot based visual debugging of GSNeighbour logic
doesnt map cleanly to GSNeighboursEx logic, which calls GSN multiple
times, thus leading to the visual debug data of multiple GSN iterations
overlapping onto the same plot.

However it does give a very very rough idea still. And also for maxTries
of 1, it will still give proper visual debug plot.

NOTE: No point in trying to work with a list/array of axes, to allow
seperate plotting of each iteration of GSN for now. If required later,
it can be implemented easily.

Also localCenters and lGroupMap purposefully not initialised outside
the loop, so that if someone calls with maxTries of 0, it will raise
a exception.