AnalPlot:GSNeighbours:bSort, _distance
Added a _distance helper function to calculate the 2D space distance
between two points in the 2D space.

Added a new argument bSort, which is True by default. This sorts the
list of localCenters identified by GSN/GSNEx.

This helps one see if a given x,y point is near or far from the origin
in a relative manner. And use it for example to give say red markers
for points near to origin and green to points farthest from origin,
as part of visualisation for example.

The document also updated a bit.