AnalPlot:GSNeighbours: Truely neighbours mode by default
NewBehaviour:
Now by default GSNeighbours assigns each data point to the local
neighbourhood it belongs too, even if the data point is towards the
edges of its neighbourhood and so potentially some other adjacent
neighbourhood's localCenter is nearer to it.

OldBehaviour:
If one wants the logic to decide the localCenters/Poles/PointsOf
Attraction, based on neighbourhood criteria (distance between
points), but finally assign a point to the group whose localCenter
is nearest to it, even if its a different neighbourhood, then
remember to set the new bLocalCenterNearest arg to True.

Chances are the new behaviour may appear better visually, but
logically it could have some other localCenter near to it. So based
on your end use case needs, decide which semantic you want.